stock_prediction/backend/requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
yfinance
tensorflow
ta-lib
ta
requests
python-dotenv
